requirements.txt
pygame
numpy
python-dotenv